Create README.md
Embarking on a 100-day coding journey focused on Data Structures and Algorithms (DSA) is both a challenging and rewarding endeavor. It’s an opportunity to push the boundaries of your problem-solving abilities, delve deep into core DSA concepts, and enhance your analytical skills. Here are a few words of encouragement